+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+- Remove functions `ts-require-language` and `ts-load-language`.
+- Added `tree-sitter-load`.
 
 ## [0.3.0] - 2020-02-21
 - Used Emacs's 1-based byte positions and line numbers instead of 0-based byte offsets and row coordinates.